+++ src/main/scala/info/gratour/common/types/DateTimeFmt.java
@@ -1,0 +1,14 @@
+/*******************************************************************************
+ *  Copyright (c) 2019, 2022 lucendar.com.
+ *  All rights reserved.
+ *
+ *  Contributors:
+ *     KwanKin Yau (alphax@vip.163.com) - initial API and implementation
+ *******************************************************************************/
+package info.gratour.common.types;
+
+public enum DateTimeFmt {
+    ISO_OFFSET_DATE_TIME,
+    CONVENIENT_DATE_TIME,
+    CONVENIENT_DATE_TIME_WITH_MILLIS;
+}