+++ src/main/scala/info/gratour/common/types/rest/RestTypes.scala
@@ -7,6 +7,8 @@
  * ******************************************************************************/
 package info.gratour.common.types.rest
 
+import info.gratour.common.error.ErrorWithCode
+
 import java.util
 
 object RestTypes {
@@ -41,8 +43,19 @@ case class Pagination(
 }
 
 object Pagination {
+  final val DEFAULT_PAGE = 1
+  final val DEFAULT_LIMIT = 20
   final val DEFAULT: Pagination = Pagination(20, 1)
   final val FIRST_ONE: Pagination = Pagination(1, 1)
+  final val MAX_LIMIT: Int = 2000
+
+  def check(page: Int, limit: Int): Unit = {
+    if (page <= 0)
+      throw ErrorWithCode.invalidParam("__page")
+
+    if (limit <= 0 || limit > MAX_LIMIT)
+      throw ErrorWithCode.invalidParam("__limit")
+  }
 
   def pageCount(totalRecordCount: Long, pageSize: Int): Long = {
     var r = totalRecordCount / pageSize